moving repository functions to service
